commit 9 01/03/19 16.00 Student+ JSP for all actions+ validation 50/50 of ID

+++ src/main/java/by/chester/servlet/MainStudentServlet.java
@@ -1,9 +1,11 @@
 package by.chester.servlet;
+
 import by.chester.dao.DaoFactory;
 import by.chester.dao.PersistException;
 import by.chester.dao.StudentDao;
 import by.chester.entities.Student;
 import by.chester.mySqlDAO.MySqlDaoFactory;
+
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -11,7 +13,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
 import java.util.List;
-
 
 @WebServlet("/main")
 public class MainStudentServlet extends HttpServlet {
@@ -31,47 +32,39 @@ public class MainStudentServlet extends HttpServlet {
             getAllStudents(req, resp);
         }
         if (answer == "Create") {
-            createStudents(req, resp);
+            createStudent(req, resp);
         }
         if (answer == "Update") {
-            updateStudents(req, resp);
+            updateStudent(req, resp);
         }
         if (answer == "Delete") {
-            deleteStudents(req, resp);
+            deleteStudent(req, resp);
         }
     }
 
-    private void deleteStudents(HttpServletRequest req, HttpServletResponse resp) {
+    private void deleteStudent(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         try {
             DaoFactory factory = new MySqlDaoFactory();
             StudentDao studentDao = factory.getMySqlStudentDao();
             Student st5 = new Student();
+            if (req.getParameter("id") != null && validationForId(Integer.parseInt(req.getParameter("id")))) {
+                st5.setId(Integer.parseInt(req.getParameter("id")));
+                studentDao.delete(st5);
+                req.setAttribute("wrongId", "correct");
+            } else {
+                req.setAttribute("wrongId", "dfghjk");
+            }
             getServletContext().getRequestDispatcher("/StudentFrameDelete.jsp").forward(req, resp);
-            try {
-                if(validationForId(Integer.parseInt(req.getParameter("id")))) {
-                    st5.setId(Integer.parseInt(req.getParameter("id")));
-                    studentDao.delete(st5);
-                    req.setAttribute("error","Unknown user");
-                } else {
-                    req.setAttribute("error","Unknown user");
-                }
-            } catch (NumberFormatException e) {
-            throw new IOException("ID должен быть числом.");
-        }
-    } catch (PersistException e) {
-            e.printStackTrace();
-        } catch (ServletException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
+        } catch (PersistException | NumberFormatException | IOException | ServletException e) {
             e.printStackTrace();
         }
     }
 
     private boolean validationForId(int id) {
-        return id % 1 == 0 && id > 0;
+        return id > 0 && id < 1000;
     }
 
-    private void updateStudents(HttpServletRequest req, HttpServletResponse resp) {
+    private void updateStudent(HttpServletRequest req, HttpServletResponse resp) {
         try {
             DaoFactory factory = new MySqlDaoFactory();
             StudentDao studentDao = factory.getMySqlStudentDao();
@@ -92,7 +85,7 @@ public class MainStudentServlet extends HttpServlet {
         }
     }
 
-    private void createStudents(HttpServletRequest req, HttpServletResponse resp) {
+    private void createStudent(HttpServletRequest req, HttpServletResponse resp) {
         try {
             DaoFactory factory = new MySqlDaoFactory();
             StudentDao studentDao = factory.getMySqlStudentDao();
@@ -115,7 +108,7 @@ public class MainStudentServlet extends HttpServlet {
     private void getAllStudents(HttpServletRequest req, HttpServletResponse resp) {
         try {
             DaoFactory factory = new MySqlDaoFactory();
-            List<Student> students =factory.getMySqlStudentDao().getAll();
+            List<Student> students = factory.getMySqlStudentDao().getAll();
             req.setAttribute("students", students);
             getServletContext().getRequestDispatcher("/student.jsp").forward(req, resp);
         } catch (PersistException e) {
@@ -127,7 +120,7 @@ public class MainStudentServlet extends HttpServlet {
         }
     }
 
-    private String checkAction (HttpServletRequest req){
+    private String checkAction(HttpServletRequest req) {
         if (req.getParameter("getAll") != null) {
             return "getAll";
         }
